inventory_service/requirements.txt
fastapi==0.104.1
strawberry-graphql[fastapi]==0.215.1
psycopg2-binary==2.9.9
sqlalchemy==1.4.53
alembic==1.12.1
pydantic==2.5.0
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
databases[postgresql]==0.8.0
aiofiles==23.2.1